+++ tests/fixtures/nuxt-global-components/components/global/Foo.vue
@@ -1,0 +1,3 @@
+<template>
+  <p>global foo</p>
+</template>