+++ tests/fixtures/nuxt-global-components/components/islands/Isle.vue
@@ -1,0 +1,3 @@
+<template>
+  <p>island</p>
+</template>